+++ views/default/widgets/amapnews_featured/content.php
@@ -41,7 +41,7 @@ if (elgg_instanceof($owner, 'user')) {
             $featured = '';
             if ($post->is_featured()) {
                 $featured_icon = elgg_view('output/img', array(
-                    'src' => elgg_get_simplecache_url('amapnews/icon/featured.png'),
+                    'src' => amapnews_getFeaturedIcon(),
                     'alt' => elgg_echo('amapnews:featured'),
                 ));    
                 $featured = elgg_format_element('div', array('style' => "float:right;"), $featured_icon);